backend/requirements.txt
flask
flask_cors
flask_sqlalchemy
graphene
graphene_sqlalchemy
lask_graphql